@@ -1,20 +1,20 @@
 {
-    "name": "ladamalina/lingua-stem-ru",
+    "name": "andrewdanilov/lingua-stem-ru",
     "description": "Стеммер Портера. Извлечение корня из русского слова, без словаря.",
     "license": "MIT",
     "authors": [
         {
-            "name": "Nadezhda Ryabtsova",
-            "email": "nadezhdaryabtsova@gmail.com"
+            "name": "Andrew Danilov",
+            "email": "andrewdanilov@gmail.com"
         }
     ],
     "minimum-stability": "dev",
     "require": {
-        "php": ">=5.3.0"
+        "php": ">=5.6.0"
     },
     "autoload": {
         "psr-4": {
-            "Stem\\": "src/"
+            "andrewdanilov\\stem\\": "src/"
         }
     },
     "require-dev": {
